core: BinaryLogProvider should only provide a BinaryLog
No behavioral changes in this PR
- Move the actual logic into BinaryLog.java, so that BinaryLogProvider
  only provides the BinaryLog
- the loading is now done via ServiceProvider util instead of rolling
  our own in BinaryLogProvider
