+++ core/src/main/java/io/grpc/internal/ManagedChannelImpl.java
@@ -524,8 +524,8 @@ final class ManagedChannelImpl extends ManagedChannel implements Instrumented<Ch
     this.transportFactory =
         new CallCredentialsApplyingTransportFactory(clientTransportFactory, this.executor);
     Channel channel = new RealChannel();
-    if (builder.binlogProvider != null) {
-      channel = builder.binlogProvider.wrapChannel(channel);
+    if (builder.binlog != null) {
+      channel = builder.binlog.wrapChannel(channel);
     }
     this.interceptorChannel = ClientInterceptors.intercept(channel, interceptors);
     this.stopwatchSupplier = checkNotNull(stopwatchSupplier, "stopwatchSupplier");
